# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Photo platform API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface LoginBody {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export type PhotoLicense = (typeof PhotoLicense)[keyof typeof PhotoLicense];

export const PhotoLicense = {
  cc0: "cc0",
  "cc-by": "cc-by",
  "cc-by-sa": "cc-by-sa",
  editorial: "editorial",
  "all-rights-reserved": "all-rights-reserved",
} as const;

export type PhotoStatus = (typeof PhotoStatus)[keyof typeof PhotoStatus];

export const PhotoStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface Photo {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  /** @nullable */
  blurHash?: string | null;
  width: number;
  height: number;
  photographerName: string;
  /** @nullable */
  photographerAvatarUrl?: string | null;
  tags: string[];
  likes: number;
  downloads: number;
  isFeatured: boolean;
  /** @nullable */
  uploadedBy?: string | null;
  createdAt: string;
  /** @nullable */
  camera?: string | null;
  /** @nullable */
  lens?: string | null;
  /** @nullable */
  aperture?: string | null;
  /** @nullable */
  shutterSpeed?: string | null;
  /** @nullable */
  iso?: number | null;
  /** @nullable */
  focalLength?: string | null;
  license: PhotoLicense;
  status: PhotoStatus;
}

export interface PhotoListResponse {
  photos: Photo[];
  total: number;
  page: number;
  limit: number;
}

export type CreatePhotoBodyLicense =
  (typeof CreatePhotoBodyLicense)[keyof typeof CreatePhotoBodyLicense];

export const CreatePhotoBodyLicense = {
  cc0: "cc0",
  "cc-by": "cc-by",
  "cc-by-sa": "cc-by-sa",
  editorial: "editorial",
  "all-rights-reserved": "all-rights-reserved",
} as const;

export type CreatePhotoBodyStatus =
  (typeof CreatePhotoBodyStatus)[keyof typeof CreatePhotoBodyStatus];

export const CreatePhotoBodyStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface CreatePhotoBody {
  title: string;
  /** @nullable */
  description?: string | null;
  imageUrl: string;
  width: number;
  height: number;
  photographerName: string;
  /** @nullable */
  photographerAvatarUrl?: string | null;
  tags: string[];
  isFeatured?: boolean;
  /** @nullable */
  camera?: string | null;
  /** @nullable */
  lens?: string | null;
  /** @nullable */
  aperture?: string | null;
  /** @nullable */
  shutterSpeed?: string | null;
  /** @nullable */
  iso?: number | null;
  /** @nullable */
  focalLength?: string | null;
  license?: CreatePhotoBodyLicense;
  status?: CreatePhotoBodyStatus;
}

export type UpdatePhotoBodyLicense =
  (typeof UpdatePhotoBodyLicense)[keyof typeof UpdatePhotoBodyLicense];

export const UpdatePhotoBodyLicense = {
  cc0: "cc0",
  "cc-by": "cc-by",
  "cc-by-sa": "cc-by-sa",
  editorial: "editorial",
  "all-rights-reserved": "all-rights-reserved",
} as const;

export type UpdatePhotoBodyStatus =
  (typeof UpdatePhotoBodyStatus)[keyof typeof UpdatePhotoBodyStatus];

export const UpdatePhotoBodyStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface UpdatePhotoBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  imageUrl?: string;
  width?: number;
  height?: number;
  photographerName?: string;
  /** @nullable */
  photographerAvatarUrl?: string | null;
  tags?: string[];
  isFeatured?: boolean;
  /** @nullable */
  camera?: string | null;
  /** @nullable */
  lens?: string | null;
  /** @nullable */
  aperture?: string | null;
  /** @nullable */
  shutterSpeed?: string | null;
  /** @nullable */
  iso?: number | null;
  /** @nullable */
  focalLength?: string | null;
  license?: UpdatePhotoBodyLicense;
  status?: UpdatePhotoBodyStatus;
}

export interface Collection {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  photoCount: number;
  createdAt: string;
}

export interface CollectionWithPhotos {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  photoCount: number;
  createdAt: string;
  photos: Photo[];
}

export interface CreateCollectionBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface AddPhotoToCollectionBody {
  photoId: number;
}

export interface Tag {
  name: string;
  photoCount: number;
}

export interface SiteSummary {
  totalPhotos: number;
  totalLikes: number;
  totalDownloads: number;
  totalCollections: number;
  totalTags: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListPhotosParams = {
  page?: number;
  limit?: number;
  /**
   * @nullable
   */
  tag?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  sort?: ListPhotosSort;
};

export type ListPhotosSort =
  (typeof ListPhotosSort)[keyof typeof ListPhotosSort];

export const ListPhotosSort = {
  latest: "latest",
  popular: "popular",
  trending: "trending",
} as const;

export type GetMyPhotosParams = {
  page?: number;
  limit?: number;
};